Rename plugin folder
