+++ packages/jbrowse-web/src/plugins/DataHubManagerDrawerWidget/components/ConfirmationDialog.js
