+++ packages/jbrowse-web/src/plugins/DataHubManagerDrawerWidget/components/DataHubDrawerWidget.js
