+++ packages/jbrowse-web/src/plugins/DataHubManagerDrawerWidget/index.js
