+++ packages/jbrowse-web/src/plugins/DataHubManagerDrawerWidget/model.js
